requirements.txt
requests
matplotlib
PyYAML